requirements.txt
Django==3.1.7
pycryptodome==3.10.1
gunicorn==20.1.0
whitenoise==5.2.0
Pillow==8.1.2
pytz==2021.1